% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibge.R
\name{ibge_aggregates}
\alias{ibge_aggregates}
\title{List IBGE aggregate tables}
\usage{
ibge_aggregates()
}
\value{
tibble: research_id, research_name, aggregate_id, aggregate_name
}
\description{
Returns all available aggregate surveys/tables from the IBGE SIDRA system.
}
